simem/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
django-filter==24.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11
